// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
      flights: {
        Row: {
          created_at: string | null
          flight_no: string | null
          id: number
          target_cg_long: number | null
        }
        Insert: {
          created_at?: string | null
          flight_no?: string | null
          id?: number
          target_cg_long?: number | null
        }
        Update: {
          created_at?: string | null
          flight_no?: string | null
          id?: number
          target_cg_long?: number | null
        }
        Relationships: []
      }
      optimization_layout: {
        Row: {
          created_at: string | null
          flight_id: number | null
          id: number
          job_id: number | null
          position_code: string | null
          uld_id: string | null
          weight: number | null
          xpos: number | null
          ypos: number | null
        }
        Insert: {
          created_at?: string | null
          flight_id?: number | null
          id?: number
          job_id?: number | null
          position_code?: string | null
          uld_id?: string | null
          weight?: number | null
          xpos?: number | null
          ypos?: number | null
        }
        Update: {
          created_at?: string | null
          flight_id?: number | null
          id?: number
          job_id?: number | null
          position_code?: string | null
          uld_id?: string | null
          weight?: number | null
          xpos?: number | null
          ypos?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "optimization_layout_flight_id_fkey"
            columns: ["flight_id"]
            isOneToOne: false
            referencedRelation: "flights"
            referencedColumns: ["id"]
          },
        ]
      }
      positions: {
        Row: {
          id: number
          position_code: string | null
          xpos: number | null
          ypos: number | null
        }
        Insert: {
          id?: number
          position_code?: string | null
          xpos?: number | null
          ypos?: number | null
        }
        Update: {
          id?: number
          position_code?: string | null
          xpos?: number | null
          ypos?: number | null
        }
        Relationships: []
      }
      uld_manifest: {
        Row: {
          flight_id: number | null
          id: number
          uld_id: string | null
          weight: number | null
        }
        Insert: {
          flight_id?: number | null
          id?: number
          uld_id?: string | null
          weight?: number | null
        }
        Update: {
          flight_id?: number | null
          id?: number
          uld_id?: string | null
          weight?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "uld_manifest_flight_id_fkey"
            columns: ["flight_id"]
            isOneToOne: false
            referencedRelation: "flights"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never